+++ include/cantera/thermo/IonsFromNeutralVPSSTP.h
@@ -391,7 +391,7 @@ public:
      * @param dlnActCoeffdlnN    Output vector of derivatives of the
      *                         log Activity Coefficients. length = m_kk * m_kk
      */
-    virtual void getdlnActCoeffdlnN(const int ld, doublereal* const dlnActCoeffdlnN) ;
+    virtual void getdlnActCoeffdlnN(const size_t ld, doublereal* const dlnActCoeffdlnN) ;
 
 
     //! Get the Salt Dissociation Coefficients
